requirements.txt
py-cord
PyNaCl
deepgram-sdk
